updateTruckDriverWorking

+++ src/main/java/pl/coderslab/endingproject/controller/TruckDriverController.java
@@ -1,5 +1,6 @@
 package pl.coderslab.endingproject.controller;
 
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import pl.coderslab.endingproject.dao.TruckDriverDao;
@@ -8,6 +9,9 @@ import pl.coderslab.endingproject.entity.TruckDriver;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.Instant;
+import java.time.LocalDate;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 @Controller
@@ -27,22 +31,21 @@ public class TruckDriverController {
         return "truckdriver/add";
     }
 
+    @RequestMapping("/add")
+    public String addTruckDriver(
+            @RequestParam String firstName,
+            @RequestParam String lastName,
+            @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate psychoTestDate,
+            @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate medTestDate,
+            @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate driverLicenseDate) {
 
-    @PostMapping("/add")
-    @ResponseBody
-    public String addTruckDriver(@RequestParam Long driverId,
-                                 @RequestParam String firstName,
-                                 @RequestParam String lastName,
-                                 @RequestParam Instant psychoTestDate,
-                                 @RequestParam Instant medTestDate,
-                                 @RequestParam Instant driverLicenseDate) {
         TruckDriver truckDriver = new TruckDriver();
-        truckDriver.setDriverId(driverId);
         truckDriver.setFirstName(firstName);
         truckDriver.setLastName(lastName);
-        truckDriver.setPsychoTestDate(psychoTestDate);
-        truckDriver.setMedTestDate(medTestDate);
-        truckDriver.setDriverLicenseDate(driverLicenseDate);
+        truckDriver.setPsychoTestDate(psychoTestDate.atStartOfDay().atZone(ZoneId.systemDefault()).toInstant());
+        truckDriver.setMedTestDate(medTestDate.atStartOfDay().atZone(ZoneId.systemDefault()).toInstant());
+        truckDriver.setDriverLicenseDate(driverLicenseDate.atStartOfDay().atZone(ZoneId.systemDefault()).toInstant());
+
         truckDriverDao.saveTruckDriver(truckDriver);
         return "redirect:/truckdriver/list";
     }
@@ -69,11 +72,14 @@ public class TruckDriverController {
         return truckDriver;
     }
 
-    @ResponseBody
     @GetMapping("/delete")
-    public String delete() {
-        TruckDriver truckDriver = truckDriverDao.getTruckDriver(1L);
-        truckDriverDao.delete(truckDriver);
-        return "Skasowano kierowcę";
+    public String deleteTruckDriver(@RequestParam Long id) {
+        TruckDriver truckDriver = truckDriverDao.getTruckDriver(id);
+        if (truckDriver != null) {
+            truckDriverDao.delete(truckDriver);
+        }
+        return "redirect:/truckdriver/list";
     }
+
+
 }
